+++ R/areaplot.R
@@ -175,6 +175,8 @@ areaplot.default <- function(x, y=NULL, prop=FALSE, rev=FALSE, add=FALSE,
   if(is.null(ylab))
     ylab <- deparse(substitute(y))
 
+  if(anyDuplicated(x))
+    stop("duplicated values on x-axis are not allowed")
   y <- as.matrix(y)
   if(length(x) != nrow(y))
     stop("'x' and 'y' lengths differ")
@@ -190,8 +192,8 @@ areaplot.default <- function(x, y=NULL, prop=FALSE, rev=FALSE, add=FALSE,
     y <- prop.table(y, 1)
   y <- t(rbind(0, apply(y, 1, cumsum)))
   na <- is.na(x) | apply(is.na(y), 1, any)
-  y <- y[!na,][order(x[!na]),]
-  x <- x[!na][order(x[!na])]
+  y <- y[!na,][order(x[!na]),]  # reorder y by x
+  x <- x[!na][order(x[!na])]    # reorder x last
 
   if(!add)
     suppressWarnings(matplot(x, y, type="n", xlab=xlab, ylab=ylab, ...))
@@ -250,7 +252,7 @@ areaplot.formula <- function(formula, data, subset, na.action, xlab=NULL,
   if(ncol(mf[-1]) == 0 || ncol(mf[-1]) >= 3)
     stop("formula must specify 1 or 2 variables after the tilde")
   if(anyDuplicated(mf[-1]))
-    stop("duplicated values after the tilde - try another formula or subset")
+    stop("duplicated values on x-axis are not allowed")
   x <- sort(unique(mf[[2]]))
   if(is.null(xlab))
     xlab <- names(mf)[2]
@@ -260,6 +262,8 @@ areaplot.formula <- function(formula, data, subset, na.action, xlab=NULL,
     if(ncol(mf[-1]) != 1)
       stop("formula with cbind() must specify 1 variable after the tilde")
     lhs <- mf[[1]]
+    if(is.null(ylab))
+      ylab <- ""
     areaplot.default(x, lhs, xlab=xlab, ylab=ylab, ...)
   }
   else